points in polygons with QuadTree
Former-commit-id: a2b549f10376e6f8a1ae7976737dc9633a32e1f9
Former-commit-id: 90af1564d74a5239cc90e1f85b50774bc51f8213
Former-commit-id: e52d59b3184a4206e4ebf8c211497150d2ac624c [formerly 488662cb8893bc15354f0c555557e08f2f8211b0]
Former-commit-id: e607306c50190af0f61d8194ed6183ad69cde40a
Former-commit-id: 4fc8ec9708e87131a62776da67d43b9ee8da41e8
Former-commit-id: a01f5f0137fa3caa0de1b51ddffb74edfe915d06

+++ src/RcppExports.cpp
@@ -197,15 +197,15 @@ BEGIN_RCPP
 END_RCPP
 }
 // point_in_polygon
-bool point_in_polygon(NumericVector vertx, NumericVector verty, float pointx, float pointy);
+bool point_in_polygon(NumericVector vertx, NumericVector verty, double pointx, double pointy);
 RcppExport SEXP _lidR_point_in_polygon(SEXP vertxSEXP, SEXP vertySEXP, SEXP pointxSEXP, SEXP pointySEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericVector >::type vertx(vertxSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type verty(vertySEXP);
-    Rcpp::traits::input_parameter< float >::type pointx(pointxSEXP);
-    Rcpp::traits::input_parameter< float >::type pointy(pointySEXP);
+    Rcpp::traits::input_parameter< double >::type pointx(pointxSEXP);
+    Rcpp::traits::input_parameter< double >::type pointy(pointySEXP);
     rcpp_result_gen = Rcpp::wrap(point_in_polygon(vertx, verty, pointx, pointy));
     return rcpp_result_gen;
 END_RCPP
